assistente-backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
google-generativeai==0.8.5
python-dotenv==1.1.0
django-cors-headers>=4.0
